+++ DatadogRUM/Sources/DataModels/RUMDataModels+objc.swift
@@ -818,6 +818,7 @@ public enum objc_RUMActionEventContainerSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -833,6 +834,7 @@ public enum objc_RUMActionEventContainerSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -846,6 +848,7 @@ public enum objc_RUMActionEventContainerSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMActionEventContainerView)
@@ -1109,6 +1112,7 @@ public enum objc_RUMActionEventSource: Int {
         case .kotlinMultiplatform?: self = .kotlinMultiplatform
         case .electron?: self = .electron
         case .rumCpp?: self = .rumCpp
+        case .maui?: self = .maui
         }
     }
 
@@ -1125,6 +1129,7 @@ public enum objc_RUMActionEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -1139,6 +1144,7 @@ public enum objc_RUMActionEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMActionEventStream)
@@ -1905,6 +1911,7 @@ public enum objc_RUMErrorEventContainerSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -1920,6 +1927,7 @@ public enum objc_RUMErrorEventContainerSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -1933,6 +1941,7 @@ public enum objc_RUMErrorEventContainerSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMErrorEventContainerView)
@@ -2793,6 +2802,7 @@ public enum objc_RUMErrorEventErrorSourceType: Int {
         case .windows?: self = .windows
         case .macos?: self = .macos
         case .linux?: self = .linux
+        case .maui?: self = .maui
         }
     }
 
@@ -2811,6 +2821,7 @@ public enum objc_RUMErrorEventErrorSourceType: Int {
         case .windows: return .windows
         case .macos: return .macos
         case .linux: return .linux
+        case .maui: return .maui
         }
     }
 
@@ -2827,6 +2838,7 @@ public enum objc_RUMErrorEventErrorSourceType: Int {
     case windows
     case macos
     case linux
+    case maui
 }
 
 @objc(DDRUMErrorEventErrorThreads)
@@ -2978,6 +2990,7 @@ public enum objc_RUMErrorEventSource: Int {
         case .kotlinMultiplatform?: self = .kotlinMultiplatform
         case .electron?: self = .electron
         case .rumCpp?: self = .rumCpp
+        case .maui?: self = .maui
         }
     }
 
@@ -2994,6 +3007,7 @@ public enum objc_RUMErrorEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -3008,6 +3022,7 @@ public enum objc_RUMErrorEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMErrorEventStream)
@@ -3754,6 +3769,7 @@ public enum objc_RUMLongTaskEventContainerSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -3769,6 +3785,7 @@ public enum objc_RUMLongTaskEventContainerSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -3782,6 +3799,7 @@ public enum objc_RUMLongTaskEventContainerSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMLongTaskEventContainerView)
@@ -4212,6 +4230,7 @@ public enum objc_RUMLongTaskEventSource: Int {
         case .kotlinMultiplatform?: self = .kotlinMultiplatform
         case .electron?: self = .electron
         case .rumCpp?: self = .rumCpp
+        case .maui?: self = .maui
         }
     }
 
@@ -4228,6 +4247,7 @@ public enum objc_RUMLongTaskEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -4242,6 +4262,7 @@ public enum objc_RUMLongTaskEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMLongTaskEventStream)
@@ -4917,6 +4938,7 @@ public enum objc_RUMResourceEventContainerSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -4932,6 +4954,7 @@ public enum objc_RUMResourceEventContainerSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -4945,6 +4968,7 @@ public enum objc_RUMResourceEventContainerSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMResourceEventContainerView)
@@ -5884,6 +5908,7 @@ public enum objc_RUMResourceEventSource: Int {
         case .kotlinMultiplatform?: self = .kotlinMultiplatform
         case .electron?: self = .electron
         case .rumCpp?: self = .rumCpp
+        case .maui?: self = .maui
         }
     }
 
@@ -5900,6 +5925,7 @@ public enum objc_RUMResourceEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -5914,6 +5940,7 @@ public enum objc_RUMResourceEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMResourceEventStream)
@@ -6720,6 +6747,7 @@ public enum objc_RUMViewEventContainerSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -6735,6 +6763,7 @@ public enum objc_RUMViewEventContainerSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -6748,6 +6777,7 @@ public enum objc_RUMViewEventContainerSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMViewEventContainerView)
@@ -7105,6 +7135,7 @@ public enum objc_RUMViewEventSource: Int {
         case .kotlinMultiplatform?: self = .kotlinMultiplatform
         case .electron?: self = .electron
         case .rumCpp?: self = .rumCpp
+        case .maui?: self = .maui
         }
     }
 
@@ -7121,6 +7152,7 @@ public enum objc_RUMViewEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -7135,6 +7167,7 @@ public enum objc_RUMViewEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMViewEventStream)
@@ -7782,6 +7815,8 @@ public enum objc_RUMViewEventViewLoadingType: Int {
         case .fragmentRedisplay?: self = .fragmentRedisplay
         case .viewControllerDisplay?: self = .viewControllerDisplay
         case .viewControllerRedisplay?: self = .viewControllerRedisplay
+        case .sessionRenewal?: self = .sessionRenewal
+        case .bfCache?: self = .bfCache
         }
     }
 
@@ -7796,6 +7831,8 @@ public enum objc_RUMViewEventViewLoadingType: Int {
         case .fragmentRedisplay: return .fragmentRedisplay
         case .viewControllerDisplay: return .viewControllerDisplay
         case .viewControllerRedisplay: return .viewControllerRedisplay
+        case .sessionRenewal: return .sessionRenewal
+        case .bfCache: return .bfCache
         }
     }
 
@@ -7808,6 +7845,8 @@ public enum objc_RUMViewEventViewLoadingType: Int {
     case fragmentRedisplay
     case viewControllerDisplay
     case viewControllerRedisplay
+    case sessionRenewal
+    case bfCache
 }
 
 @objc(DDRUMViewEventViewLongTask)
@@ -8631,6 +8670,7 @@ public enum objc_RUMViewUpdateEventContainerSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -8646,6 +8686,7 @@ public enum objc_RUMViewUpdateEventContainerSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -8659,6 +8700,7 @@ public enum objc_RUMViewUpdateEventContainerSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMViewUpdateEventContainerView)
@@ -9016,6 +9058,7 @@ public enum objc_RUMViewUpdateEventSource: Int {
         case .kotlinMultiplatform?: self = .kotlinMultiplatform
         case .electron?: self = .electron
         case .rumCpp?: self = .rumCpp
+        case .maui?: self = .maui
         }
     }
 
@@ -9032,6 +9075,7 @@ public enum objc_RUMViewUpdateEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -9046,6 +9090,7 @@ public enum objc_RUMViewUpdateEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMViewUpdateEventStream)
@@ -9693,6 +9738,8 @@ public enum objc_RUMViewUpdateEventViewLoadingType: Int {
         case .fragmentRedisplay?: self = .fragmentRedisplay
         case .viewControllerDisplay?: self = .viewControllerDisplay
         case .viewControllerRedisplay?: self = .viewControllerRedisplay
+        case .sessionRenewal?: self = .sessionRenewal
+        case .bfCache?: self = .bfCache
         }
     }
 
@@ -9707,6 +9754,8 @@ public enum objc_RUMViewUpdateEventViewLoadingType: Int {
         case .fragmentRedisplay: return .fragmentRedisplay
         case .viewControllerDisplay: return .viewControllerDisplay
         case .viewControllerRedisplay: return .viewControllerRedisplay
+        case .sessionRenewal: return .sessionRenewal
+        case .bfCache: return .bfCache
         }
     }
 
@@ -9719,6 +9768,8 @@ public enum objc_RUMViewUpdateEventViewLoadingType: Int {
     case fragmentRedisplay
     case viewControllerDisplay
     case viewControllerRedisplay
+    case sessionRenewal
+    case bfCache
 }
 
 @objc(DDRUMViewUpdateEventViewLongTask)
@@ -10617,6 +10668,7 @@ public enum objc_RUMVitalAppLaunchEventContainerSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -10632,6 +10684,7 @@ public enum objc_RUMVitalAppLaunchEventContainerSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -10645,6 +10698,7 @@ public enum objc_RUMVitalAppLaunchEventContainerSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMVitalAppLaunchEventContainerView)
@@ -10908,6 +10962,7 @@ public enum objc_RUMVitalAppLaunchEventSource: Int {
         case .kotlinMultiplatform?: self = .kotlinMultiplatform
         case .electron?: self = .electron
         case .rumCpp?: self = .rumCpp
+        case .maui?: self = .maui
         }
     }
 
@@ -10924,6 +10979,7 @@ public enum objc_RUMVitalAppLaunchEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -10938,6 +10994,7 @@ public enum objc_RUMVitalAppLaunchEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMVitalAppLaunchEventStream)
@@ -11724,6 +11781,7 @@ public enum objc_RUMVitalDurationEventContainerSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -11739,6 +11797,7 @@ public enum objc_RUMVitalDurationEventContainerSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -11752,6 +11811,7 @@ public enum objc_RUMVitalDurationEventContainerSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMVitalDurationEventContainerView)
@@ -12015,6 +12075,7 @@ public enum objc_RUMVitalDurationEventSource: Int {
         case .kotlinMultiplatform?: self = .kotlinMultiplatform
         case .electron?: self = .electron
         case .rumCpp?: self = .rumCpp
+        case .maui?: self = .maui
         }
     }
 
@@ -12031,6 +12092,7 @@ public enum objc_RUMVitalDurationEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -12045,6 +12107,7 @@ public enum objc_RUMVitalDurationEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMVitalDurationEventStream)
@@ -12770,6 +12833,7 @@ public enum objc_RUMVitalOperationStepEventContainerSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -12785,6 +12849,7 @@ public enum objc_RUMVitalOperationStepEventContainerSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -12798,6 +12863,7 @@ public enum objc_RUMVitalOperationStepEventContainerSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMVitalOperationStepEventContainerView)
@@ -13061,6 +13127,7 @@ public enum objc_RUMVitalOperationStepEventSource: Int {
         case .kotlinMultiplatform?: self = .kotlinMultiplatform
         case .electron?: self = .electron
         case .rumCpp?: self = .rumCpp
+        case .maui?: self = .maui
         }
     }
 
@@ -13077,6 +13144,7 @@ public enum objc_RUMVitalOperationStepEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -13091,6 +13159,7 @@ public enum objc_RUMVitalOperationStepEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDRUMVitalOperationStepEventStream)
@@ -13444,6 +13513,7 @@ public enum objc_TelemetryConfigurationEventSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -13458,6 +13528,7 @@ public enum objc_TelemetryConfigurationEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -13470,6 +13541,7 @@ public enum objc_TelemetryConfigurationEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDTelemetryConfigurationEventTelemetry)
@@ -13598,6 +13670,11 @@ public class objc_TelemetryConfigurationEventTelemetryConfiguration: NSObject {
 
     public var isMainProcess: NSNumber? {
         root.swiftModel.telemetry.configuration.isMainProcess as NSNumber?
+    }
+
+    public var mauiVersion: String? {
+        set { root.swiftModel.telemetry.configuration.mauiVersion = newValue }
+        get { root.swiftModel.telemetry.configuration.mauiVersion }
     }
 
     public var mobileVitalsUpdatePeriod: NSNumber? {
@@ -13824,6 +13901,11 @@ public class objc_TelemetryConfigurationEventTelemetryConfiguration: NSObject {
     public var trackNetworkRequests: NSNumber? {
         set { root.swiftModel.telemetry.configuration.trackNetworkRequests = newValue?.boolValue }
         get { root.swiftModel.telemetry.configuration.trackNetworkRequests as NSNumber? }
+    }
+
+    public var trackResourceHeaders: objc_TelemetryConfigurationEventTelemetryConfigurationTrackResourceHeaders {
+        set { root.swiftModel.telemetry.configuration.trackResourceHeaders = newValue.toSwift }
+        get { .init(swift: root.swiftModel.telemetry.configuration.trackResourceHeaders) }
     }
 
     public var trackResources: NSNumber? {
@@ -14123,6 +14205,30 @@ public enum objc_TelemetryConfigurationEventTelemetryConfigurationTrackFeatureFl
     case longTask
 }
 
+@objc(DDTelemetryConfigurationEventTelemetryConfigurationTrackResourceHeaders)
+@_spi(objc)
+public enum objc_TelemetryConfigurationEventTelemetryConfigurationTrackResourceHeaders: Int {
+    internal init(swift: TelemetryConfigurationEvent.Telemetry.Configuration.TrackResourceHeaders?) {
+        switch swift {
+        case nil: self = .none
+        case .defaultHeaders?: self = .defaultHeaders
+        case .custom?: self = .custom
+        }
+    }
+
+    internal var toSwift: TelemetryConfigurationEvent.Telemetry.Configuration.TrackResourceHeaders? {
+        switch self {
+        case .none: return nil
+        case .defaultHeaders: return .defaultHeaders
+        case .custom: return .custom
+        }
+    }
+
+    case none
+    case defaultHeaders
+    case custom
+}
+
 @objc(DDTelemetryConfigurationEventTelemetryConfigurationTrackingConsent)
 @_spi(objc)
 public enum objc_TelemetryConfigurationEventTelemetryConfigurationTrackingConsent: Int {
@@ -14391,6 +14497,7 @@ public enum objc_TelemetryDebugEventSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -14405,6 +14512,7 @@ public enum objc_TelemetryDebugEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -14417,6 +14525,7 @@ public enum objc_TelemetryDebugEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDTelemetryDebugEventTelemetry)
@@ -14666,6 +14775,7 @@ public enum objc_TelemetryErrorEventSource: Int {
         case .kotlinMultiplatform: self = .kotlinMultiplatform
         case .electron: self = .electron
         case .rumCpp: self = .rumCpp
+        case .maui: self = .maui
         }
     }
 
@@ -14680,6 +14790,7 @@ public enum objc_TelemetryErrorEventSource: Int {
         case .kotlinMultiplatform: return .kotlinMultiplatform
         case .electron: return .electron
         case .rumCpp: return .rumCpp
+        case .maui: return .maui
         }
     }
 
@@ -14692,6 +14803,7 @@ public enum objc_TelemetryErrorEventSource: Int {
     case kotlinMultiplatform
     case electron
     case rumCpp
+    case maui
 }
 
 @objc(DDTelemetryErrorEventTelemetry)
@@ -14828,4 +14940,4 @@ public class objc_TelemetryErrorEventView: NSObject {
 
 // swiftlint:enable force_unwrapping
 
-// Generated from https://github.com/DataDog/rum-events-format/tree/ea41a41f19117b04ed9a06977fdeb8f7a90319e3
+// Generated from https://github.com/DataDog/rum-events-format/tree/ed69a908b5f05a97b984498526d50c0e97284c06