+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Unreleased
 
+# 3.11.0 / 12-05-2026
+
+- [IMPROVEMENT] Add support for `maui` source for cross-platform RUM events from .NET MAUI applications. See [#2891][]
+- [FEATURE] Add client state management to `DatadogFlags` module. See [#2719][]
+- [IMPROVEMENT] Skip malformed RUM attributes individually instead of dropping the entire event, and log clear error messages. See [#2844][]
+- [FIX] Propagate native `anonymous_id` to WebView RUM and Log events. See [#2847][]
+- [IMPROVEMENT] Add device properties `isLowRam`, `logicalCpuCount`, `totalRam` to the `LogEvent` Objective-C API. See [#2854][]
+- [IMPROVEMENT] Add `collectAccessibility` property to the Objective-C API of RUM configuration. See [#2855][]
+- [IMPROVEMENT] Add Objective-C APIs for `networkSettledResourcePredicate` and `nextViewActionPredicate` of RUM configuration. See [#2856][]
+- [FIX] Fix `us2_fed` intake hostname. See [#2866][]
+- [FEATURE] Add `redactedStatusCodes` to `Trace.Configuration.URLSessionTracking` to configure which HTTP status codes have their resource URL replaced with the status code string in spans. Defaults to `[404]` for backward compatibility. See [#2845][]
+
 # 3.10.0 / 16-04-2026
 
 - [FEATURE] Add Mobile Heatmaps support. See [#2829][]
@@ -1116,6 +1128,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2773]: https://github.com/DataDog/dd-sdk-ios/pull/2773
 [#2776]: https://github.com/DataDog/dd-sdk-ios/pull/2776
 [#2791]: https://github.com/DataDog/dd-sdk-ios/pull/2791
+[#2719]: https://github.com/DataDog/dd-sdk-ios/pull/2719
 [#2794]: https://github.com/DataDog/dd-sdk-ios/pull/2794
 [#2802]: https://github.com/DataDog/dd-sdk-ios/pull/2802
 [#2807]: https://github.com/DataDog/dd-sdk-ios/pull/2807
@@ -1123,6 +1136,14 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2817]: https://github.com/DataDog/dd-sdk-ios/pull/2817
 [#2827]: https://github.com/DataDog/dd-sdk-ios/pull/2827
 [#2829]: https://github.com/DataDog/dd-sdk-ios/pull/2829
+[#2844]: https://github.com/DataDog/dd-sdk-ios/pull/2844
+[#2847]: https://github.com/DataDog/dd-sdk-ios/pull/2847
+[#2854]: https://github.com/DataDog/dd-sdk-ios/pull/2854
+[#2845]: https://github.com/DataDog/dd-sdk-ios/pull/2845
+[#2855]: https://github.com/DataDog/dd-sdk-ios/pull/2855
+[#2856]: https://github.com/DataDog/dd-sdk-ios/pull/2856
+[#2866]: https://github.com/DataDog/dd-sdk-ios/pull/2866
+[#2891]: https://github.com/DataDog/dd-sdk-ios/pull/2891
 
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin